langgraph-data-analyst/requirements.txt
langgraph==1.1.3
langchain==1.2.13
langchain-core==1.2.23
google-genai==1.69.0
pandas==3.0.1
numpy==2.4.4
matplotlib==3.10.8
seaborn==0.13.2
duckdb==1.5.1
fastapi==0.135.2
uvicorn==0.42.0
pydantic==2.12.5
pydantic-settings==2.9.1
python-dotenv==1.2.2
loguru==0.7.3
